pyproject.toml
[tool.poetry]
name = "numblitz-server"
version = "0.1.0"
description = ""
authors = ["ShanThatos <shanthkoka@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
psutil = "^5.9.6"
python-dotenv = "^1.0.0"
authlib = "^1.2.1"
itsdangerous = "^2.1.2"
httpx = "^0.25.0"
sqla-wrapper = "^6.0.0"
psycopg = {extras = ["binary"], version = "^3.1.12"}
pyjwt = "^2.8.0"
mathgen = {git = "https://github.com/ShanThatos/mathgen.git"}
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"


[tool.poetry.group.dev.dependencies]
black = "^23.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py312"]